requirements.txt
cachelib==0.9.0
cffi==1.15.1
click==8.1.3
colorama==0.4.5
cryptography==37.0.4
DateTime==4.5
Flask==2.1.3
Flask-Cors==3.0.10
Flask-Session==0.4.0
Flask-SQLAlchemy==2.5.1
greenlet==1.1.2
gunicorn==20.1.0
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
pycparser==2.21
PyJWT==2.4.0
pytz==2022.1
six==1.16.0
SQLAlchemy==1.4.39
uuid==1.30
Werkzeug==2.2.1
zipp==3.8.1
zope.interface==5.4.0
pytest_flask~=1.2.0
pytest~=7.1.2